Remove content export option
