+++ src/javascript/AdminPanel/Export.utils.js
@@ -1,0 +1,10 @@
+export const exportJSONFile = (data, fileName) => {
+    const jsonString = JSON.stringify(data, null, 2);
+    const blob = new Blob([jsonString], {type: 'application/json;charset=utf-8;'});
+    const link = document.createElement('a');
+    link.href = URL.createObjectURL(blob);
+    link.setAttribute('download', `${fileName}.json`);
+    document.body.appendChild(link);
+    link.click();
+    document.body.removeChild(link);
+};
