--- src/javascript/ExportContent/ExportContent.utils.js
@@ -1,8 +1,0 @@
-export const extractAndFormatContentTypeData = data => {
-    const nodes = data?.jcr?.nodeTypes?.nodes || [];
-    return nodes.map(n => ({
-        label: n.displayName || n.name,
-        value: n.name,
-        iconStart: n.icon
-    }));
-};
